vsi-lldb: process launch status while attaching.
The launch status might change after it first became "RunningGame".
E.g. when binaries are missing, the status might be first "RunningGame"
for a short period of time, but then it becomes "GameLaunchEnded".

Thus, we need to check the status not only before we are trying to
attach to the game, but also while we attaching and interrupt the
attach process as soon as we see that the game is not in the running
state.

GitOrigin-RevId: fadaed5edb3efae59eeaf2041dc9a9818fa5ebd9