vsi-lldb: Fix computation of module address in minidumps
When replacing a placeholder module with a real module in minidump,
we compute the address of the object file from the difference
between file address and and file offset in the first code section.
However, the LLVM 10 toolchain can create segments with different
differences of file address-offset. This CL just replaces the
calculation with direct call to SBModule.GetObjectFileHeaderAddress.

GitOrigin-RevId: 5d4429c4ca50674579983319fba9ef90071c0165