vsi-lldb: show account name when suggesting to stop the game.
Only one game can be launched for each developer / test account.
This CL modifies the prompt that suggests users to stop the game
when they already have another launch. It adds information about
the type of account that is used and the account itself (email
for dev account, stadia gamer name for test account).

GitOrigin-RevId: ee3e1636a0a103e68127cdd2c829f207d05a430c