vsi-lldb: fix attach flow for the new launch API.
This CL fixes the attach and the attach to core flows
for the launch API.

We only have a reference to IVsiGameLaunch when we launched
the game. This is not a case for the crash dump debugging.

For the attach to the running game flow, we don't have an
option to verify match the game launch with PID using the
launch API. We also don't have information which account
(test or dev) has been used to launch the game. Thus for
the attach flow we don't check the launch status at the
moment.

GitOrigin-RevId: 64f95ab455bdb7af814bce79d2b372812e86757f