[grpc-server] Rename anonymous namespaces in stack traces
This replaces (anonymous namespace) with `anonymous namespace' in
stack traces. A frame with a function name that starts with left paren
apparently crashes Visual Studio when the parallel stacks window
is open, so we really want to avoid leading left parens. The new
name for anonymous namespace was chosen to be consistent with
native VS.

GitOrigin-RevId: 81143ad0e002b87e5d622d450e337f2381f1887b