Reland "[grpc-server] Remove repeated param list from call stack"
For C++ methods/function, LLDB returns parameter list as part of
the function name (SBFrame::GetFunctionName). This CL removes
the parameter list from the function name in function info
(if the parameter list is present) and in frame's function name.

As a result, the Call Stack window in Visual Studio does not show
those silly repeated parameter lists anymore.

GitOrigin-RevId: 9a69bf6a8e7193de8fcdb1b84f96b989a745b8de