[build][project-system] Extract ProjectSystem dependencies in version-specific libraries
The extension depends on `Microsoft.VisualStudio.ProjectSystem`, which has breaking changes in API in Visual Studio 2019. Currently we have multiple build configurations and have to build different VSIX for different Visual Studio versions.

In this change ProjectSystem dependencies are extracted in version-specific libraries, that are loaded dynamically in runtime. This way we can have only one build configuration and one VSIX that is compatible with both Visual Studio 2017 and 2019.

Along with the refactoring all conditional references (i.e. <When>) are removed, effectively making the build identical for (Debug|Release)2017 and (Debug|Release)2019 configurations. The configurations are not yet removed, because other parts of the infrastructure rely on it.

So for now we still build two different VSIX, but the only difference between them is the manifest files. The contents (assemblies, references, dependencies, etc.) are exactly the same.

GitOrigin-RevId: 1a1e7d24aa40730d0f2cbb89f56e37f88eda31c5