[natvis] Make copies of VariableInformation holding context variables
If a Natvis expression is equal to a variable name, LLDB or lldb-eval
may return a reference to the variable. Since display value of these
expressions are determined after all expression are evaluated, display
values of all expressions corresponding to the same variable will be
equal to the last value of the variable. The issue arises when the
variable's value is changed between two such expressions, which is
allowed for custom Natvis variables (<Variable> rule in
<CustomListItems>).

Consider the following example:

<CustomListItems>
  <Variable Name="counter" InitialValue="0" />
  <Loop>
    <Break Condition="counter == 3" />
    <Exec>counter++</Exec>
    <Item>counter</Item>
  </Loop>
</CustomListItems>

The expected output is [1, 2, 3], but due to deferred evaluations
of display values, the output is [3, 3, 3].

This CL copies evaluations results of custom Natvis variables (aka.
context variables) in order to prevent the issue.

GitOrigin-RevId: f9dbb1ec97aa4e8217fde967cc5b0005abc130f4