[debugger] Enable side effects for frame evaluations
Allow side effects for the expressions evaluated in the frame context: Watch
Window and Immediate Window.

This is implemented by allowing side effects for all "frame" evaluations.
Technically, we should do this only for the Immediate Window (and maybe
Watch Window), but this is basically the only case where frame
evaluations are used.

GitOrigin-RevId: 9ce05ec93ccf21e6ebf4a0152453c7c4c1711d8a