[grpc-server] Set working directory for DebuggerGrpcServer
Use the directory with the binary as working directory. Normally it
doesn't matter, because the server doesn't load any files via relative
paths and the dependent DLLs are available via PATH and PYTHONPATH.
However when debugging or just running the extension from Visual Studio
the working directory is set to the output build directory. The working
directory has precedence for loading DLLs, so liblldb.dll is picked up
from the build output directory, NOT the deployed extension.

GitOrigin-RevId: f3b2adbc1a57f85d4907e928e3387c53e822d94c