[disassembly] Fix moving instruction pointer in disassembly window
Moving the instruction pointer in the disassembly window didn't work
because for that it is required to check whether the destination
address lies within the same function.

Code context created in the disassembly stream didn't contain the
proper user name and used address instead.

This change passes the function name to the context. In order to
avoid many remote calls to grpc server when opening disassembly
window, it is doing the lookup lazily.

GitOrigin-RevId: 66f1319378da5fe6f2e439f89a1e963835ed58ca