[debugger] Load project Natvis files after attaching logger
Logs from LoadProjectFiles() were missed because
natvisLogListener.OnNatvisLogEvent was not hooked up yet when the
method was called. This CL moves LoadProjectFiles() to a slightly
later point.

The logs show up in the Output window in the Natvis pane when the
Natvis log level is set to e.g. Verbose, see
Tools > Options > Stadia > Natvis diagnostic messages.

GitOrigin-RevId: 9b4c8683b9f7c00cb0b6c2e74808a9a57b98998e