[vsi-lldb][launch-api] Fix invalid test account error
When any player endpoint is selected, the test account is ignored.
So we need to prevent test account loading early on, because
otherwise the specified test account is used to select gamelets
and it may result in game launch fail when the test account is
Invalid. Also test_account query parameter can not be supported,
because many test account related verifications are done before
query string parsing.

GitOrigin-RevId: 0d8244eeb6a6fa29c3f772fa481d283b50b3309c