[vsi-lldb][launch-api] Use simplified overrides for player endpoints
Simplified overrides provide a simple api to configure game client as
opposed to complex configuration via combination of display, system,
audio and device settings. Using the old overrides produces similar
code in all LaunchGame callers.
Enabling it only for player endpoints first, because simplified
overrides are still new and TestClient was for now the only way
to launch a game.

GitOrigin-RevId: 3b364f22498d8e0b2de256482dd9f4834e7f186f