[VSFake] Implement call stack and threads window
Implements a view for threads and call stacks.

The call stacks window gets updated automatically whenever the
selected thread changes. Since it watches for the SelectedThread and
queues a new job to refresh the stack frames, ProgramStoppedJob has to
queue a sub-job to set the program execution state.

The threads window has to be refreshed manually.

GitOrigin-RevId: 181353267377942737a27cb459635a4f0e4adaa9