[GameletMountChecker] Adjust invalid mount detection
The current mount detection assumes that the workstation directory
is mounted to /mnt/developer and that it uses fuse.sshfs. The former
is no longer the case in current gamelets, where the mount setup has
been changed so that /mnt/developer is always the instance SSD.
The latter is not true for AS30, where it is fuse.cdc_fuse_fs.

This CL adjusts the mount detection to work for both AS20 and AS30.
Warnings about invalid mount setups are shown now if /mnt/developer is
not overlaid on /srv/game/assets and
a) the binaries are deployed to /mnt/developer or
b) asset streaming is deactivated
In both cases, the binaries built by Visual Studio can't be used by
the game.

The SshTunnels functionality to detect the mount point for AS20 is no
longer required with the new mount setup. The functionality detected
cases where the target path was streamed, so that copying binaries
from the target path /mnt/developer would be streamed back to the
target path, causing corruption. This can no longer happen with the
new mount setup since data is never streamed back to the workstation.
It is a valid case now to stream from the target path AND to deploy
binaries to /mnt/developer.

The GetDevices* unit tests are removed since they do not add coverage.
The code paths are already hit by the other tests.

GitOrigin-RevId: 5caf1ed7185b40dcb43c0a0fd01fe7fdf5cad5e7