[launch-api] Add enable orbit option to debugging tab
Similar to the "Enable Dive" option, the "Enable Orbit" option
enables enhanced profiling capabilites for the Orbit profiler, such
command buffer timings, for the application being run on the
instance.

In contrast to the Dive, RGP, and RenderDoc options, this option is
not necessary to profile an application using Orbit, but adds
additional features. Therefore, this change also slightly adjusts
the wording of the other options.

GitOrigin-RevId: 2a1d4f084300b4dd8170f15a5935b054ecd9f95d