[DebugEngine] Fix preflight binary check
The preflight binary check looks up the game binary in /mnt/developer.
This does not work when the game is streamed from a local workstation
directory. In that case, the workstation directory is mounted to
/mnt/workstation and mirrored to /srv/game/assets.

In this CL, the check is changed to use /srv/game/assets. This is the
source of truth for the game, so it's the proper choice.

GitOrigin-RevId: 1cc600d62b22b93574af78e6e4c5ab718a0cbcab