[VSFake] Handle exception events
This CL handles events of type IDebugExceptionEvent2. It terminates
the program and sets an exception that bubbles up to the corresponding
RunUntil() call, which will then fail the test with a proper error
message. The exception is cleared after being thrown, so that the
test's TearDown() can proceed without error.

GitOrigin-RevId: d4646a5af1eaf14e4b25e4c07ce9c18ffec42b2f