[VSFake] Allow setting pass count before program runs
This CL move the infra to set a breakpoint pass count from Breakpoint
to BreakpointsWindow and store pass count and style if the breakpoint
is not bound yet (i.e. if the program is not started yet). Once the
program is started, the CL sets the pass count and style. This allows
adding breakpoints and setting the pass count before a program is
started.

Also contains
- a method to delete a single breakpoint,
- better naming for GenericJobs and
- some styling fixes.

GitOrigin-RevId: 0217e42b858fa24461cca45894c5c44652e6ccc1