[vsix] Migrate YetiVSIPackage to AsyncPackage
This change allows to load Stadia vsi extension even when no
Stadia project is opened (for example to debug a minidump).
If the ProvideAutoLoad attributes are used without migration to
AsyncPackage first, VS will block the extension.
Also removes obsolete unit-tests. We don't have to validate that
VS is able to register custom commands.

GitOrigin-RevId: 33973d57bebe77cf5f47f9d3ad3a0958586ae7bc