[YetiVSI] Allow Catching SyncAsync's Exceptions
Exceptions thrown in event handlers can not be caught (easily).
Uses of RemoteFile's SyncAsync method typical surround the "await call"
with a try/catch block. However, the "ProcessException" thrown in the
event handler will not be catched by this.

In this change, we store the error strings in the event handler and
throw the exception afterwards instead of throwing it in the handler.

This fixes a crash of VS when starting a process with "Orbit options"
enabled and no Orbit installed.

GitOrigin-RevId: ef2d2aee680d40efe69ca4c2c038a841b601905e