[launch-api][DebugEngine] Add property to customize the launch path
Prior to having this property, the VSI would always launch the program
using the relative file name of the executable, which always expands to
/mnt/developer/<TargetFileName>.

The new property user property "GgpGameletLaunchExecutable" allows the
user to customize this. This property can be either a relative or
absolute path. Relative paths are relative to '/srv/assets/game'.

If this property is empty, then VSI falls back to the original relative
target file name for backward compatibility.

GitOrigin-RevId: 6ed7d18c1f13ce142207272e6bf5629297a7fa42