[debugger] Use ELFSharp lib to parse ELF files
At the moment in order to get BuildId or a symbol file
link location we use llvm-objdump.exe to get the section
of interest and manually parse the response. This CL introduces
ModuleParser, which uses ElfSharp instead, and since we now
don't need a separate process for each module, the symbol loading
gets significantly faster (under 1s vs ~minute before the change).
ModuleParser.ParseRemoteBuildIdInfoAsync is a copy of
ElfFileUtil.ReadBuildIdAsync with SSHTarget being set
(needed for checking BuildId of a binary on a gamelet). There
will be a couple of calls to this function so it's fine to
leave it as is, at least for now (the calls to IFileUtil will
be replaced with IModuleParser in the following CLs).

GitOrigin-RevId: a295fc4c6511797558e211cff143d0ae0bcbb879