[debugger] Replace IFileUtil with IModuleParser in SymbolLoader
Use ModuleParser instance to parse modules (symbol file name and
location). There is no need to handle exceptions since the code pre-
validates different types of errors and sets the Error message in
the output accordingly.
IBinaryFileUtil's ReadSymbolFileNameAsync and ReadSymbolFileDirAsync
are not used anymore, therefore are removed from the code.

GitOrigin-RevId: b5488d4f81d484c08a9fb31f24498c42ce5a9480