[natvis] Improve Natvis std::string workaround for LLVM 11+
When compiled with -fno-standalone-debug, clang-10 (and possibly
earlier versions) used to omit debug info required to display
std::string properly. As a workaround, piece of Natvis code was added
that relies on c_str() method in the cases where it's available.
Since clang-11 enough information is produced by the compiler and
the workaround is no longer needed (it's also significantly slower).

This change keeps the original workaround if the debugged binary was
compiled with Clang version 10 or less, and it removes it otherwise.

GitOrigin-RevId: f9c93d4f71f75cb059abf3ab0df437ae6d1674af