[Game Launch] Refactor ChromeClientsLauncher
Pushes the decision whether to launch Chrome and which URL to use to
ChromeClientsLauncher. This gets rid of some duplicate code in
the ChromeClientLauncher app and the debug engine.

Also gets rid of a call in VsiGameLaunch that was basically just
calling a method on a passed-in object. The (one) caller now just
calls the method directly. This gets rid of some interface methods and
unit tests that just replicate ChromeClientsLauncher tests.

GitOrigin-RevId: 785822fa690b49adee2c29126aaf13723d1bc627