[Game Launch] Launch Orbit from GgpDebugQueryTarget
Launches Orbit if the flag combo
  DebugLaunchOptions.NoDebug | DebugLaunchOptions.Profiling
is passed into GgpDebugQueryTarget. Right now, this code path is not
exercised yet, but in a future CL a command will be added that
triggers it.

Also note that Orbit right now does not handle the passed command line
args properly. It expects pid@gameletId, but we pass path@gameletId
because that's more efficient and simpler on our side.

GitOrigin-RevId: 450440c3af4fb3f3ea389a2bd1a9ac30f5cbe424