[YetiVSI Project] Fix DeployPythonAndToolchainDependencies
Fixes the DeployPythonAndToolchainDependencies flag to speed up builds
by excluding Python and LLDB from the VSIX file and using a local
version instead. There was some ordering issue in the build that
prevented that the proper Python path was written to the
local_python_dir.txt file.

Setting DeployPythonAndToolchainDependencies to false effectively saves
10s-15s iteration time.

GitOrigin-RevId: 7129a554db5116465ba46ca009647b08ad90c3fb