[debugger] Split binaries and symbols loading into 2 separate steps
- Replace
```
foreach module in modules:
  checkModuleName(module);
  module?=loadBinary(module);
  loadSymbols(module);
```
with
```
filteredByName = FilterByName(modules);
notPlaceholders = loadBinariesIfNeeded(filteredByName);
loadSymbolsIfNeeded(notPlaceholders);
```
It also allows to get rid of extra checks (for example module name
cannot be null in BinaryLoader or SymbolLoader). The unit-tests
were fixed accordingly.

- IModuleFileLoadMetricsRecorder accepts a LoadSymbolData instance
as a parameter to record to avoid keeping initial `modules` array
during the whole processing time

- try-finally blocks are not needed, they were guarding against losing
the searchLog on the task cancellation requested in the middle of
the processing block. Now, if the task is cancelled, it will happen
before searchLog was populated, and there is no need to attempt to
save it

- ModuleSearchLogHolder contains mapping of module's ID -> search log.

GitOrigin-RevId: c766a6fe5d17b4e0c3fe2ad53bb4113125b1a5eb