[Game Launch] Refactor Orbit command class
Refactors the LaunchWithOrbitCommand class into a more generic
LaunchWithProfilerCommand class that can be used for Dive as well.

So which DebugLaunchOption will be used to indicate a Dive launch?
The plan is to use DebugLaunchOptions.DetachOnStop, since that does
not make sense in combination with DebugLaunchOptions.NoDebug (which
is always passed for profiler launches), so it should never interfere
with real use cases.

GitOrigin-RevId: c6113e1e58807ad623f5f02cdc3e36e3ba790812