[DebugEngine] Stop the debug session in case of error.
Previously, in some cases of errors (for example, if there was no Stadia instance),
we did not stop the debug session.
In this change I refactored LaunchSuspended and Attach method to always
stop debugging session if error occured.
I also added cancellation logic to the LLDB debugger launching task
in order to stop the operation in case of errors or cancelation.

GitOrigin-RevId: cab4b4b5085835cbb3087e3b15e5c46cf921c5ef