[YetiVSI.Tests] Fix coverage report generation
If YetiVSI.Tests is run as a startup project, it generates a code
coverage report. This is currently broken because $(OutputBaseDir) is
not set.

This CL replaces it by "..\..\", which works because the
test is executed with cwd set to bin/$(Configuration). Note that
$(ProjectDir) is not available for SDK-style projects. Which is weird,
but I guess that's the way it is. The CL also keeps the window open in
case something fails.

The coverage report ends up in YetiVSI.Tests/bin/$(Configuration)/coverage.

GitOrigin-RevId: 575b8a066e0030bf7e2b4194441808b1880a98cf