[debugger] Enable setting PC to an arbirarty address
The main change in this CL is using `SBFrame::SetPC(addr)` to move the
execution pointer instead of resolving the line number and using
`JumpToLine`. This enables jumping to an arbitrary address.

Simplify the checks in `CanSetNextStatement` -- do the sanity check for
the current thread and then compare PC/function name. If the function
names are different, return `S_FALSE`, which will make Visual Studio to
show a dialog "Are you sure you want to jump to a different function?".

`CanSetNextStatement` is no longer called from `SetNextStatement`, since
Visual Studio will always call it beforehand (and also handle the case
of jumping to a different function).

Several other things are fixed along the way:

* Move the function name resolution logic inside the `DebugCodeContext`,
  because the name needs to be re-resolved after the address changes
  (e.g. after `Add/Subtract)`.

* Merge `DebugMemoryContext` into `DebugCodeContext`. They're basically
  the same thing and from the implementation perspective it doesn't make
  a difference if they're two different classes.

GitOrigin-RevId: 3de7467b2879a7ceb113fae4b3031bf21043942e