[Profiling] Add infra to run port forwarding tunnels
Dive, RGP and Renderdoc expect SSH port forwarding processes to be
present when the game starts. Right now, these processes are started
from YetiDebugTransport. This code is only run when a debugger is
attached. However, it is not a good idea to profile with debugger
attached, as that deteriorates performance. Hence, the port forwarding
processes should be started whenever the game is started, whether a
debugger is attached or not.

This CL is prep work for that. It adds
- SshTunnelProcess: Wraps an SSH port forwarding process.
- GameLifetimeWatcher: Triggers a callback when the game shuts down.
When the game shuts down, the tunnels are shut down as well.

SshTunnelManager is the overall manager class that coordinates
spinning up and shutting down tunnels.

Also moves the ManagedProcess.OnExit call to after logging, as the
handlers might dispose the process, so ExitCode isn't valid anymore.

This CL does not affect current behavior, the new code is not yet
hooked up.

GitOrigin-RevId: ca4627d6473daf5a5906f4b77b37d8a89219066d