[Game Launch] Run profiler SSH processes earlier
Currently, the SSH port forwarding processes for Dive, RGP and
Renderdoc are started from YetiDebugTransport, which requires a
debugger to be attached. This does not make sense for profiling,
though.

This CL uses the previously developed infrastructure around
SshTunnelManager to run the port forwarding processes from
GgpDebugQueryTarget instead. The processes are stopped automatically
when the game exits. This way, profilers can be used even if the user
hits CTRL + F5 (run without debugger).

GitOrigin-RevId: 0bb98237ad193160dcc4a7cd1e0bce04fa1641a7