[Game Launch] Add UI controls to launch Dive
Changes the button to launch Dive to a dropdown to launch any
profiler. Also adds menu items for the three additional profiler.

Renderdoc and RGP might be added in future updates, but they might
require some more work. Renderdoc is more complicated than just
starting a binary. RGP requires an explicit capture step (either from
button in Test Client or ggp rgb capture command), and RGP isn't
installed with the SDK, so we don't know the location of the tool and
can't start it.

GitOrigin-RevId: 8ed09d31c3d2d31d5e8fc41819bfec0df9e63fe8