[natvis] Implement gRPC for expression compilation
This CL is a part of efforts to support Priority attribute.
Expressions in Natvis are evaluated as encountered. With this change
it will be possible to pre-compile Natvis expressions during
visualizer selection process. The current idea is to use it for
properly selecting the visualizer, but in the future this could also
be used to re-use compiled expressions while evaluating Natvis.

GitOrigin-RevId: c238520553e837c0828bf2e8709f4b3663ef7a11