[Game Launch] Kill previous Orbit instance
If a game is launched with Orbit as profiler, and there is an old
instance of Orbit still open, terminate that instance. Orbit was
designed for only one instance, and multiple instances can lead to
undefined behavior. Florian said it's OK to just kill it, Orbit saves
all data.

Dive, in contrast, does work fine with multiple instances.

GitOrigin-RevId: c128f3acefd4bef78be4e5d60ced1b647ad2a7a0