[port-supplier] Fix deadlock in PortSupplier ctor
Visual Studio might deadlock forever when using "Attach to Process" menu
command.

This happens because `DebugPortSupplier` (which is created when `Attach
to Process` is invoked) tries to get an instance of `YetiVSIService`,
which triggers initilization of the `YetiVSIPackage` (if it hasn't been
loaded yet). If the package has already been loaded, then everything
works fine. If the package is not loaded yet, then it's loaded and
initialized, but it deadlocks when calling
`serviceManager.GetJoinableTaskContext()` on the call to
`GetGlobalService(typeof(SComponentModel))`.

It seems that using global `ThreadHelper.JoinableTaskContext` instead of
getting one from the `SComponentModel` avoid the deadlock.

GitOrigin-RevId: a0764ff5d64411d9e9ffebcb2518aab1dad43c24