[natvis] Disable Priority attribute by default
Priority attribute and compilation of Natvis are related features
that were recently supported in our engine. The idea was to implement
them under an experimental flag until they're better tested.

However, only the compilation feature was submitted under the
experimental flag, leaving the Priority attribute enabled by default.
The impact of this isn't critical, but it cause minor differences in
our tests.

To fix this, Priority attribute is disabled by default in this CL.

GitOrigin-RevId: 5432db0e05f1e9d0b0f11bd09fa41157657a7b11