[natvis] Handle compilation for "lldb-eval with fallback"
Before this change, "lldb-eval" and "lldb-eval with fallback" options
were handled the same way. "lldb-eval with fallback" uses LLDB if
lldb-eval fails in certain cases during evaluation. In this change
the behavior was change in a way that if "NotImplemented" error was
returned by lldb-eval during compilation, compilation isn't considered
failed since LLDB will likely handle such cases successfully.

There are few exceptions to this rule: "NotImplemented" errors in
<HeadPointer> and <Variable> elements result in failed compilation
even if the "lldb-eval with fallback" option is used. The reason is
that results of these elements affect compilation of other elements.

GitOrigin-RevId: 1d9056bdab48986ae80e1ee6abd5029dfbfd08fc