[YetiVSI] Fix process ID retrieval when attaching
- Fix the retrieval of the process ID when we're attaching to a process
given that the Build ID is not set. When attaching to a process when the
Build ID is not set, we expect a warning suggesting to set the build ID
but we should be able to attach to the running process. This was failing
as the process ID wasn't retrieved correctly.
- Renamed GetRemoteProcessIdAndCheckExecutable to CheckExecutable and
updated the signature of the method.

GitOrigin-RevId: 53199079be9150700d34b4b99b4d2dfc17a6980c