[Debugger] Check build IDs in GgpDebugQueryTarget
Modifies the startup flow so that local and remote binary build IDs
are compared in GgpDebugQueryTarget, after deploying the binary.
Also detects and uses the overlay directories from the mount
configuration to figure out where the binary might be.

This fixes a race condition where the binary might not be visible in
/srv/game/assets until the folder is remounted, which happens during
game start, concurrently with the previous check for build IDs.

It also fixes the issue that build IDs were never checked in
run-without-debugger (CTRL+F5) flow.

PreflightBinaryChecker should be moved outside of DebugEngine now.
This isn't done here to keep the diff smaller.

GitOrigin-RevId: 85443a177f7bbe59e14ea12d6c5440389c0dd2e7