[Debugger] Make binary-not-found critical
A PreflightBinaryCheckerException is currently thrown if the game
binary is not found. This exception was treated like a warning.
However, it should be an error.

Similar PreflightBinaryCheckerExceptions thrown if the build ids don't
match should remain a warning, though.

This CL adds a flag to PreflightBinaryCheckerException to denote
whether the exception should be treated as a warning or as an error.

GitOrigin-RevId: da6d53177b5d104fdee966b4da604d7439c2a268