[Debugger] Add temp safety net for future mount changes
If /srv/game/assets was changed from an overlay to something else
(like a bind-mount), the VSI would not be able to start up anymore
since mountConfig.OverlayDirs would be empty, so that we would not
be able to detect a binary and every attempt to debug a game would
end with "game binary not found".

This CL falls back to checking /srv/game/assets/binary if OverlayDirs
is empty. This is the right thing to do in case of a bind mount, and
it matches the old behavior of the code.

GitOrigin-RevId: 956174bff7f92dcb521ba31b40242abd8f6060be