[build][project-system] Remove YetiVSI.ProjectSystem hack
YetiVSI.ProjectSystem.v15/16/17 exist to workaround the problem of backwards incompatibility in Visual Studio APIs. In order to build a single VSIX that is compatible with both VS2017 and VS2019 (that's where the compatibility is broken), we build two versions of `YetiVSI.ProjectSystem` and then resolve the correct one in runtime. This has a benefit of resolving everything statically in the code and having tests for each version of the API, but it complicates and slows down the build.

Since the incompatibility surface is really small (a single member changed its type), we can just workaround that difference via reflection and remove the whole `YetiVSI.ProjectSystem` machinery.

GitOrigin-RevId: e145900cd7f99c59eaf75ab68ac253fb04620007