[build] Copy LLDBEval.dll with DebuggerGrpcServer
Native DLLs are not copied to the output directory automatically, that's by design in MSBuild project system. The dependency chain is as follows:
  DebuggerGrpcServer -> LLDBWorker -> LLDBEval

LLDBWorker is a C++/CLI library and thus copied to the output directory with all its managed dependencies. LLDBEval is a native library, so it needs to be copied manually.

Higher level projects (e.g. YetiVSI) should not deal with `LLDBEval.dll`, for them all dependencies should be resolved automatically.

GitOrigin-RevId: ba84ba1470f5707957bf9e258d1d46b150f5c409