[Debugger] Fix game output missing from debug pane
When a game is run with debugger attached, we mirror the game's output
on Visual Studio's Debug output pane. However, we were often missing
the last few lines, which are often the more important ones.

The reason is interesting. We were switching to the main thread on
every output line. When a game exits, we shut down all processes and
actually wait for a second for the tail job to finish. However, this
is done synchronously, so it preempts line printing tasks, so that
further event handlers for output lines are never called.

This CL queues up log lines and flushes them every 500ms. This is
better for performance, anyway.

GitOrigin-RevId: c298266ae7bd1f2dce2272acc2f197020ce09155